Set defaults for sampleSizes and embeddingSize
- Change type of sampleSize to long in LayerConfig
- Avoid specifying the defaults in tests/benchmark

+++ alpha/alpha-embeddings/src/main/java/org/neo4j/gds/embeddings/graphsage/Layer.java
@@ -25,7 +25,7 @@ import org.neo4j.gds.embeddings.graphsage.ddl4j.functions.Weights;
 import java.util.List;
 
 public interface Layer {
-    int sampleSize();
+    long sampleSize();
 
     Aggregator aggregator();
