+++ alpha/alpha-embeddings/src/main/java/org/neo4j/gds/embeddings/graphsage/LayerConfig.java
@@ -27,7 +27,7 @@ import java.util.Optional;
 public interface LayerConfig {
     int rows();
     int cols();
-    int sampleSize();
+    long sampleSize();
 
     Optional<Integer> bias();
